requirements.txt
opencv-python-headless==4.8.0.74
mediapipe==0.10.2
numpy==1.24.3
Flask~=3.0.2
itsdangerous~=2.1.2
WTForms~=3.0.1
matplotlib~=3.7.2
sklearn~=0.0.post7
scikit-learn==1.3.0
streamlit==1.32.0
pillow==10.0.0